requirements.txt
Flask==2.3.3
google-generativeai
psycopg2-binary
python-dotenv==1.0.0
gunicorn